+++ src/main/java/kr/co/mz/sns/config/security/CustomUserDetailService.java
@@ -2,7 +2,7 @@ package kr.co.mz.sns.config.security;
 
 import java.util.List;
 import kr.co.mz.sns.dto.user.GenericUserDto;
-import kr.co.mz.sns.repository.UserRepository;
+import kr.co.mz.sns.repository.user.UserRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
@@ -13,20 +13,20 @@ import org.springframework.stereotype.Service;
 @Service
 public class CustomUserDetailService implements UserDetailsService {
 
-    private final UserRepository userRepository;
-    private final ModelMapper modelMapper;
+  private final UserRepository userRepository;
+  private final ModelMapper modelMapper;
 
-    @Autowired
-    public CustomUserDetailService(UserRepository userRepository, ModelMapper modelMapper) {
-        this.userRepository = userRepository;
-        this.modelMapper = modelMapper;
-    }
+  @Autowired
+  public CustomUserDetailService(UserRepository userRepository, ModelMapper modelMapper) {
+    this.userRepository = userRepository;
+    this.modelMapper = modelMapper;
+  }
 
-    @Override
-    public CustomUserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        var userEntity = userRepository.findByEmail(email)
-            .orElseThrow(() -> new UsernameNotFoundException("Username not found"));
-        return new CustomUserDetails(modelMapper.map(userEntity, GenericUserDto.class),
-            List.of(new SimpleGrantedAuthority(userEntity.getRole())));
-    }
+  @Override
+  public CustomUserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+    var userEntity = userRepository.findByEmail(email)
+        .orElseThrow(() -> new UsernameNotFoundException("Username not found"));
+    return new CustomUserDetails(modelMapper.map(userEntity, GenericUserDto.class),
+        List.of(new SimpleGrantedAuthority(userEntity.getRole())));
+  }
 }
