+++ src/main/java/kr/co/mz/sns/controller/post/PostController.java
@@ -7,8 +7,8 @@ import kr.co.mz.sns.dto.post.PostLikeDto;
 import kr.co.mz.sns.dto.post.SelectPostDto;
 import kr.co.mz.sns.file.FileStorageService;
 import kr.co.mz.sns.service.FileService;
-import kr.co.mz.sns.service.PostLikeService;
-import kr.co.mz.sns.service.PostService;
+import kr.co.mz.sns.service.post.PostLikeService;
+import kr.co.mz.sns.service.post.PostService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -30,11 +30,11 @@ import org.springframework.web.multipart.MultipartFile;
 @RequestMapping("/api/auth/posts")
 public class PostController {
 
-    private final PostService postService;
-    private final PostLikeService postLikeService;
-    private final FileService fIleService;
+  private final PostService postService;
+  private final PostLikeService postLikeService;
+  private final FileService fIleService;
 
-    //    @RestController
+  //    @RestController
 //    public class PostController {
 //
 //        @PostMapping("/create-post")
@@ -50,43 +50,43 @@ public class PostController {
 //            return ResponseEntity.ok("Post created successfully.");
 //        }
 //    }
-    @PostMapping
-    public ResponseEntity<SelectPostDto> write(@RequestPart("files") List<MultipartFile> files,
-        @Valid @RequestParam String content) {
-        var createdDto = postService.insert(new SelectPostDto(content));
-        FileStorageService.saveFile(files);
-        return ResponseEntity
-            .status(HttpStatus.CREATED)
-            .body(
-                createdDto
-            );
-    }
-
-    @DeleteMapping("/{seq}")
-    public ResponseEntity<String> delete(@NotNull @PathVariable Long seq) {
-        postService.deleteBySeq(seq);
-        return ResponseEntity.ok(
-            "Post with ID " + seq + " has been successfully deleted"
+  @PostMapping
+  public ResponseEntity<SelectPostDto> write(@RequestPart("files") List<MultipartFile> files,
+      @Valid @RequestParam String content) {
+    var createdDto = postService.insert(new SelectPostDto(content));
+    FileStorageService.saveFile(files);
+    return ResponseEntity
+        .status(HttpStatus.CREATED)
+        .body(
+            createdDto
         );
-    }
+  }
 
-    @PutMapping("/{seq}")
-    public ResponseEntity<SelectPostDto> update(@NotNull @PathVariable Long seq,
-        @Valid @RequestBody SelectPostDto selectPostDto) {
-        return ResponseEntity.ok(
-            postService.updateBySeq(seq, selectPostDto)
+  @DeleteMapping("/{seq}")
+  public ResponseEntity<String> delete(@NotNull @PathVariable Long seq) {
+    postService.deleteBySeq(seq);
+    return ResponseEntity.ok(
+        "Post with ID " + seq + " has been successfully deleted"
+    );
+  }
+
+  @PutMapping("/{seq}")
+  public ResponseEntity<SelectPostDto> update(@NotNull @PathVariable Long seq,
+      @Valid @RequestBody SelectPostDto selectPostDto) {
+    return ResponseEntity.ok(
+        postService.updateBySeq(seq, selectPostDto)
+    );
+  }
+
+  @PostMapping("/{seq}/like")
+  public ResponseEntity<List<PostLikeDto>> like(@NotNull @PathVariable Long seq) {
+    var insertedPostLikeDto = postService.like(seq);
+    log.debug("PostLike inserted:: {}", insertedPostLikeDto);
+    return ResponseEntity
+        .status(HttpStatus.CREATED)
+        .body(
+            postLikeService.findAll(seq)
         );
-    }
-
-    @PostMapping("/{seq}/like")
-    public ResponseEntity<List<PostLikeDto>> like(@NotNull @PathVariable Long seq) {
-        var insertedPostLikeDto = postService.like(seq);
-        log.debug("PostLike inserted:: {}", insertedPostLikeDto);
-        return ResponseEntity
-            .status(HttpStatus.CREATED)
-            .body(
-                postLikeService.findAll(seq)
-            );
-    }
+  }
 
 }
