+++ src/main/java/kr/co/mz/sns/dto/post/SelectPostDto.java
@@ -4,22 +4,21 @@ import jakarta.validation.constraints.NotEmpty;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
-import kr.co.mz.sns.entity.CommentEntity;
+import kr.co.mz.sns.entity.comment.CommentEntity;
 import lombok.Data;
 
 @Data
 public class SelectPostDto {
 
-    public SelectPostDto(String content) {
-        this.content = content;
-    }
-
-    private Long seq;
-    @NotEmpty
-    private String content;
-    private Integer likes = 0;
-    private LocalDateTime createdAt;
-    private LocalDateTime modifiedAt;
-    private Long createBy;
-    private List<CommentEntity> comments = new ArrayList<>();
+  private Long seq;
+  @NotEmpty
+  private String content;
+  private Integer likes = 0;
+  private LocalDateTime createdAt;
+  private LocalDateTime modifiedAt;
+  private Long createBy;
+  private List<CommentEntity> comments = new ArrayList<>();
+  public SelectPostDto(String content) {
+    this.content = content;
+  }
 }
