--- src/main/java/kr/co/mz/sns/entity/UserEntity.java
@@ -1,44 +1,0 @@
-package kr.co.mz.sns.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@Entity
-@EntityListeners(AuditingEntityListener.class)
-@Table(name = "User")
-@Data
-@NoArgsConstructor
-public class UserEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "seq")
-    private Long seq;
-    @Column(name = "email", nullable = false)
-    private String email;
-    @Column(name = "name", nullable = false)
-    private String name;
-    @Column(name = "password", nullable = false)
-    private String password;
-    @Column(name = "role", nullable = false)
-    private String role;
-    @LastModifiedDate
-    @Column(name = "last_login_at")
-    private LocalDateTime lastLoginAt;
-    @CreatedDate
-    @Column(name = "created_at", nullable = false)
-    private LocalDateTime createdAt;
-    @Column(name = "modified_at")
-    private LocalDateTime modifiedAt;
-}
