+++ src/main/java/kr/co/mz/sns/entity/comment/CommentEntity.java
@@ -1,4 +1,4 @@
-package kr.co.mz.sns.entity;
+package kr.co.mz.sns.entity.comment;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -23,24 +23,24 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 @NoArgsConstructor
 public class CommentEntity {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "seq")
-    private Long seq;
-    @Column(name = "content", nullable = false)
-    private String content;
-    @CreatedBy
-    @LastModifiedBy
-    @Column(name = "create_by", nullable = false)
-    private Long createBy;
-    @Column(name = "post_seq", nullable = false)
-    private Long postSeq;
-    @Column(name = "comment_like", nullable = false)
-    private Long commentLike;
-    @CreatedDate
-    @Column(name = "created_at", nullable = false)
-    private LocalDateTime createdAt;
-    @LastModifiedDate
-    @Column(name = "modified_at")
-    private LocalDateTime modifiedAt;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column(name = "seq")
+  private Long seq;
+  @Column(name = "content", nullable = false)
+  private String content;
+  @CreatedBy
+  @LastModifiedBy
+  @Column(name = "create_by", nullable = false)
+  private Long createBy;
+  @Column(name = "post_seq", nullable = false)
+  private Long postSeq;
+  @Column(name = "comment_like", nullable = false)
+  private Long commentLike;
+  @CreatedDate
+  @Column(name = "created_at", nullable = false)
+  private LocalDateTime createdAt;
+  @LastModifiedDate
+  @Column(name = "modified_at")
+  private LocalDateTime modifiedAt;
 }
