+++ src/main/java/kr/co/mz/sns/repository/comment/CommentLikeRepository.java
@@ -1,7 +1,6 @@
-package kr.co.mz.sns.repository;
+package kr.co.mz.sns.repository.comment;
 
-import kr.co.mz.sns.entity.CommentEntity;
-import kr.co.mz.sns.entity.CommentLikeEntity;
+import kr.co.mz.sns.entity.comment.CommentLikeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentLikeRepository extends JpaRepository<CommentLikeEntity, Long> {
