+++ src/main/java/kr/co/mz/sns/repository/post/PostFileRepository.java
@@ -1,6 +1,6 @@
-package kr.co.mz.sns.repository;
+package kr.co.mz.sns.repository.post;
 
-import kr.co.mz.sns.entity.PostFileEntity;
+import kr.co.mz.sns.entity.post.PostFileEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostFileRepository extends JpaRepository<PostFileEntity, Long> {
