+++ src/main/java/kr/co/mz/sns/repository/post/PostLikeRepository.java
@@ -1,10 +1,10 @@
-package kr.co.mz.sns.repository;
+package kr.co.mz.sns.repository.post;
 
 import java.util.List;
-import kr.co.mz.sns.entity.PostLikeEntity;
+import kr.co.mz.sns.entity.post.PostLikeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostLikeRepository extends JpaRepository<PostLikeEntity, Long> {
 
-    List<PostLikeEntity> findByPostSeq(Long postSeq);
+  List<PostLikeEntity> findByPostSeq(Long postSeq);
 }
