+++ src/main/java/kr/co/mz/sns/repository/user/UserDetailFileRepository.java
@@ -1,10 +1,10 @@
-package kr.co.mz.sns.repository;
+package kr.co.mz.sns.repository.user;
 
-import kr.co.mz.sns.entity.UserDetailFileEntity;
+import kr.co.mz.sns.entity.user.UserDetailFileEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserDetailFileRepository extends JpaRepository<UserDetailFileEntity, Long> {
 
-    UserDetailFileEntity findByName(String name);
+  UserDetailFileEntity findByName(String name);
 
 }
