+++ src/main/java/kr/co/mz/sns/repository/user/UserDetailRepository.java
@@ -1,7 +1,7 @@
-package kr.co.mz.sns.repository;
+package kr.co.mz.sns.repository.user;
 
 import java.util.Optional;
-import kr.co.mz.sns.entity.UserDetailEntity;
+import kr.co.mz.sns.entity.user.UserDetailEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserDetailRepository extends JpaRepository<UserDetailEntity, Long> {
