+++ src/main/java/kr/co/mz/sns/service/comment/CommentLikeService.java
@@ -1,4 +1,4 @@
-package kr.co.mz.sns.service;
+package kr.co.mz.sns.service.comment;
 
 public class CommentLikeService {
