+++ src/test/java/kr/co/mz/sns/repository/UserRepositoryTest.java
@@ -3,6 +3,7 @@ package kr.co.mz.sns.repository;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import kr.co.mz.sns.repository.user.UserRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
